New protocol hook is the only protocol hook
