+++ src/protocol-hook.js
@@ -2,8 +2,13 @@ import _ from 'lodash';
 import url from 'url';
 import fs from 'fs';
 import btoa from 'btoa';
+import mime from 'mime-types';
+import pify from 'pify';
+
+require('./regenerator');
 
 const magicWords = "__magic__file__to__help__electron__compile.js";
+const fsp = pify(require('fs'));
 
 let protocol = null;
 
@@ -33,6 +38,25 @@ export function rigHtmlDocumentToInitializeElectronCompile(doc) {
   return lines.join("\n");
 }
 
+function requestFileJob(filePath, finish) {
+  fs.readFile(filePath, (err, buf) => {
+    if (err) { 
+      if (err.errno === 34) {
+        finish(-6); // net::ERR_FILE_NOT_FOUND
+        return;
+      } else {
+        finish(-2); // net::FAILED
+        return;
+      }
+    }
+    
+    finish({
+      data: buf,
+      mimeType: mime.lookup(filePath) || 'text/plain'
+    });
+  });
+}
+
 export function initializeProtocolHook(availableCompilers, initializeOpts) {
   protocol = protocol || require('protocol');
 
@@ -46,28 +70,25 @@ export function initializeProtocolHook(availableCompilers, initializeOpts) {
   let electronCompileSetupCode = initializeOpts.production ?
     `if (window.require && !window.__electron_compile_set_up) { window.__electron_compile_set_up = true; var opts = JSON.parse(decodeURIComponent(atob("${encodedOpts}"))); require('electron-compile').initForProduction(opts.cacheDir, opts.compilerInformation); }` :
     `if (window.require && !window.__electron_compile_set_up) { window.__electron_compile_set_up = true; var opts = JSON.parse(decodeURIComponent(atob("${encodedOpts}"))); require('electron-compile').initWithOptions(opts); }`;
-
-  let handler = (request) => {
+    
+  protocol.interceptBufferProtocol('file', async function(request, finish) {
     let uri = url.parse(request.url);
 
     if (request.url.indexOf(magicWords) > -1) {
-      return new protocol.RequestStringJob({
+      finish({
         mimeType: 'text/javascript',
-        data: electronCompileSetupCode
+        data: new Buffer(electronCompileSetupCode, 'utf8')
       });
+      
+      return;
     }
 
     // This is a protocol-relative URL that has gone pear-shaped in Electron,
     // let's rewrite it
     if (uri.host && uri.host.length > 1) {
-      if (!protocol.RequestHttpJob) {
-        console.log("Tried to correct protocol-relative URL, but this requires Electron 0.28.2 or higher: " + request.url);
-        return new protocol.RequestErrorJob(404);
-      }
-
-      return new protocol.RequestHttpJob({
-        url: request.url.replace(/^file:/, "https:")
-      });
+      let newUri = request.url.replace(/^file:/, "https:");
+      // TODO: Jump off this bridge later
+      finish(-2);
     }
 
     let filePath = decodeURIComponent(uri.pathname);
@@ -79,7 +100,8 @@ export function initializeProtocolHook(availableCompilers, initializeOpts) {
 
     // NB: Special-case files coming from atom.asar or node_modules
     if (filePath.match(/[\/\\]atom.asar/) || filePath.match(/[\/\\]node_modules/)) {
-      return new protocol.RequestFileJob(filePath);
+      requestFileJob(filePath, finish);
+      return;
     }
 
     let sourceCode = null;
@@ -89,53 +111,47 @@ export function initializeProtocolHook(availableCompilers, initializeOpts) {
       compiler = _.find(availableCompilers, (x) => x.shouldCompileFile(filePath));
 
       if (!compiler) {
-        return new protocol.RequestFileJob(filePath);
+        requestFileJob(filePath, finish);
       }
     } catch (e) {
       console.log(`Failed to find compiler: ${e.message}\n${e.stack}`);
-      return new protocol.RequestErrorJob(-2); // net::FAILED
+      finish(-2); // net::FAILED
+      return;
     }
 
     try {
-      sourceCode = sourceCode || fs.readFileSync(filePath, 'utf8');
+      sourceCode = sourceCode || await fsp.readFile(filePath, {encoding: 'utf8'});
     } catch (e) {
       // TODO: Actually come correct with these error codes
-      if (e.errno === 34) {
-        return new protocol.RequestErrorJob(6); // net::ERR_FILE_NOT_FOUND
+      if (e.errno === 34 /*ENOENT*/) {
+        finish(-6); // net::ERR_FILE_NOT_FOUND
+        return;
       }
 
       console.log(`Failed to read file: ${e.message}\n${e.stack}`);
-      return new protocol.RequestErrorJob(2); // net::FAILED
+      finish(-2); // net::FAILED
+      return;
     }
 
     let realSourceCode = null;
     try {
       realSourceCode = compiler.loadFile(null, filePath, true, sourceCode);
     } catch (e) {
-      return new protocol.RequestStringJob({
+      finish({
         mimeType: compiler.getMimeType(),
-        data: `Failed to compile ${filePath}: ${e.message}\n${e.stack}`
+        data: new Buffer(`Failed to compile ${filePath}: ${e.message}\n${e.stack}`)
       });
+      
+      return;
     }
 
     if (!disableAutoRendererSetup && filePath.match(/\.html?$/i)) {
       realSourceCode = rigHtmlDocumentToInitializeElectronCompile(realSourceCode, initializeOpts.cacheDir);
     }
-
-    return new protocol.RequestStringJob({
-      mimeType: compiler.getMimeType(),
-      data: realSourceCode,
+    
+    finish({
+      data: new Buffer(realSourceCode),
+      mimeType: compiler.getMimeType()
     });
-  };
-
-  // NB: Electron 0.30.4 and higher require us to call interceptProtocol, not 
-  // registerProtocol
-  let versions = _.map(process.versions['electron'].split('.'), (x) => parseInt(x));
-  let useIntercept = (versions[1] * 100 + versions[2] >= 3004);
-
-  if (useIntercept) {
-    protocol.interceptProtocol('file', handler);
-  } else {
-    protocol.registerProtocol('file', handler);
-  }
+  });
 }
